Add files via upload
In the realm of Python's code, we embark on a journey through the data symphony of Spotify and Deezer, modern virtuosos of the music streaming industry. Inspired by Plato's timeless words, our mission is to decode the recent trends shaping songs, artists, and playlists, revealing insights that empower creators and align with the evolving rhythms of listener preferences. Our analytical baton seeks to uncover hidden harmonies, emerging genres, and collaborative crescendos, offering a digital compass for artists navigating the dynamic landscape of music consumption. In this short and sweet exploration, we blend the artistry of music with the precision of Python to illuminate the ever-evolving symphony of streaming.